Use np.float64 instead of np.float in test
The np.float is deprecated and an alias for the builtin 'float'.
Since we want to test that numpy types can be serialized, we
instead use np.float64.

+++ tests/test_save_load.py
@@ -88,7 +88,7 @@ def test_number(check_save_load):  # pylint: disable=redefined-outer-name
     """
     Test number serialization.
     """
-    x = [3, np.float(2.3), 1 + 3j]
+    x = [3, np.float64(2.3), 1 + 3j]
     check_save_load(x)
